<fix> reorganize code
