--- virtual-kubelet/java/pod/let/converter.go
@@ -1,40 +1,0 @@
-/**
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package let
-
-import (
-	"strings"
-
-	"github.com/koupleless/arkctl/common/fileutil"
-	"github.com/koupleless/arkctl/v1/service/ark"
-	corev1 "k8s.io/api/core/v1"
-)
-
-type Converter struct {
-}
-
-func (*Converter) GetBizNameAndVersion(podName string) (string, string) {
-	splitted := strings.Split(podName, "-")
-	return splitted[0], strings.Join(splitted[1:], "-")
-}
-
-func (c *Converter) GetBizModel(pod *corev1.Pod) *ark.BizModel {
-	bizname, version := c.GetBizNameAndVersion(pod.Name)
-	return &ark.BizModel{
-		BizName:    bizname,
-		BizVersion: version,
-		BizUrl:     fileutil.FileUrl(pod.Labels[LABEL_KOUPLELESS_MODULE_URL]),
-	}
-}
